requirements.txt
python-dotenv
pymongo
curl_cffi
selectolax